// src/api/generated/model/signalFeedbackIn.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * LeadRadar API
 * OpenAPI spec version: 0.1.0
 */
import type { SignalFeedbackInVerdict } from './signalFeedbackInVerdict';

export interface SignalFeedbackIn {
  verdict: SignalFeedbackInVerdict;
  service_id?: string | null;
  reason?: string | null;
}